Repair weakEquality bug.
A bug was encountered where the shortening of the second tree single-child nodes was not being performed.

The consequences of this bug were that some weakly equivalent trees were being reported as being NOT WEAKLY EQUIVALENT. Therefore, weakEquality returned false when sometimes it should haver returned true.
It was never the case that weakEquality returned true when two trees were actually not weakly equivalent. Therefore, this did not have an impact on our results. In fact, whenever we expected weak equality,
this function returned correctly. This happened because the detected tree was almost always shorter (especially for non-ping based algorithms) than the source tree, and this means that the bug was not trigerred.